% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgw_download.R
\name{imgw_download}
\alias{imgw_download}
\title{Funkcja pobiera wybrane dane meteorologiczne z serwerów Instytutu Meteorologii i Gospodarki Wodnej}
\usage{
imgw_download(typ, kod, rok, path)
}
\arguments{
\item{typ}{character}

\item{kod}{character}

\item{rok}{integer}

\item{path}{character}
}
\value{

}
\description{
Funkcja pobiera wybrane dane meteorologiczne z serwerów Instytutu Meteorologii i Gospodarki Wodnej
}
\examples{
\dontrun{
 imgw_download(typ = "dobowe",
              kod = 105,
              rok = '2012',
              path = "")
}
}
