% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgw_plot.R
\name{imgw_plot}
\alias{imgw_plot}
\title{Funkcja wizualizuje wybrane dane w postaci wykresu}
\usage{
imgw_plot(typ, dane, miesiac, zmienna, stacja, dzien)
}
\arguments{
\item{typ}{character}

\item{dane}{character}

\item{miesiac}{integer}

\item{zmienna}{character}

\item{stacja}{character}

\item{dzien}{integer}
}
\value{

}
\description{
Funkcja wizualizuje wybrane dane w postaci wykresu
}
\examples{
\dontrun{
imgw_plot(typ = "miesieczne",
          dane = '',
          #miesiac = 5,
          zmienna = "cisnienie",
          stacja = "Zakopane")
}
}
