% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgw_import.R
\name{imgw_import}
\alias{imgw_import}
\title{Funckja importuje pobrane dane do środowiska R}
\usage{
imgw_import(typ, rok, kod, format, path)
}
\arguments{
\item{typ}{character}

\item{rok}{integer}

\item{kod}{character}

\item{format}{character}

\item{path}{character}
}
\value{
dataframe
}
\description{
Funckja importuje pobrane dane do środowiska R
}
\examples{
\dontrun{
Miesieczne_smt <-imgw_import(typ = "dobowe",
                             rok =  2012,
                             kod = 105,
                             format = "sdt",
                             path = "")
}
}
