% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Kody}
\alias{Kody}
\title{Kody stacji wykorzystywane do pobierania danych}
\format{
Ramka danych zawiera:
\describe{
\item{KOD}{Unikatowy kod stacji}
\item{Nazwa stacji}{Nazwa danej stacji}
}
}
\usage{

}
\description{
Kody stacji wykorzystywane do pobierania danych
}
\examples{
\dontrun{
 Kody

}
}
\keyword{datasets}
