% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kody}
\alias{kody}
\title{Kody stacji wykorzystywane do pobierania danych}
\format{
Ramka danych zawiera:
\describe{
\item{KOD}{Unikatowy kod stacji}
\item{Stacja}{Nazwa danej stacji}
}
}
\usage{
data(kody)
}
\description{
Kody stacji wykorzystywane do pobierania danych
}
\examples{
\dontrun{
 kody
}
}
\keyword{datasets}
