% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_kody.R
\docType{data}
\name{kody}
\alias{kody}
\title{Kody stacji meteorologicznych}
\format{
Ramka danych zawiera:
\describe{
\item{Kod}{numeric unikatowy kod stacji}
\item{Stacja}{character, nazwa stacji, do które przypisany jest dany kod}
}
}
\source{
Instytut Meteorologii i Gospodarki Wodnej.
\url{https://danepubliczne.imgw.pl/}
}
\usage{
kody
}
\description{
Kody stacji meteorologicznych
}
\examples{
\dontrun{
 kody
}
}
\keyword{datasets}
